HTTP Manager & Packages
Created an HTTPManager class that request the uri for the web service
we are calling data from.

Created a parser package in which will be where we will be creating our
classes to parse the data that we are retrieving from the API.

Created a model package in which will have our POJO’s.